Updated trial scripts
Replaced the original scripts with the updated ones.

Added some comments in all of them guessing what some functions do.

Added an else-condition to Q1 to allow the function to return false when the slotId was already empty.

Made a small correction in Q2 where it seemed resultId was accidentally called result when being used.

Renamed the function in Q3 based on my understanding of what it was doing because the name wasn't descriptive enough. Also made the player variable local to match the party variable and moved Player(membername) into its own variable because it's used multiple times.

Added includes and the std namespace to Q4 to load the necessary files in preprocessing. Simplified the section that searched for the player by only using the loadByPlayerName function so we don't look for the player twice.